requirements.txt
streamlit
pillow
numpy<2
scikit-learn